+++ packages/blockchain-wallet-v4-frontend/src/scenes/SecurityCenter/WalletRecoveryPhrase/RecordBackupPhrase/ThirdStep/template.js
@@ -2,11 +2,13 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import styled from 'styled-components'
 import { FormattedMessage } from 'react-intl'
-import { reduxForm } from 'redux-form'
+import { reduxForm, Field } from 'redux-form'
+import { toLower, equals } from 'ramda'
+
 import { Button, Link, Icon, Text } from 'blockchain-info-components'
-import { Form } from 'components/Form'
-import WordInput from './WordInput'
+import { Form, TextBox } from 'components/Form'
 import { SuccessOverlay } from 'components/Security'
+import { required } from 'services/FormHelper'
 
 const Container = styled.div`
   display: flex;
@@ -31,8 +33,44 @@ const VerificationContainer = styled.div`
   opacity: ${props => props.authType ? 0.3 : 1};
 `
 
+const WordContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-start;
+  align-items: center;
+  height: 60px;
+  width: 100px;
+  div:first-of-type {
+    margin-bottom: 30px;
+  }
+`
+const WordError = styled.label`
+  position: absolute;
+  top: ${props => props.errorBottom ? '35px' : '60px'};
+  right: 0;
+  display: block;
+  height: 15px;
+  font-size: 12px;
+  font-weight: 300;
+  color: ${props => props.theme['error']};
+`
+
+const validateWord = index => (value, allValues, props) => {
+  console.info(value, allValues, props, index)
+  return equals(toLower(value), toLower(props.recoveryPhrase[index])) ? undefined : <WordError errorBottom={props.errorBottom}>Invalid</WordError>
+}
+
+const languageHelper = (num) => {
+  switch (num) {
+    case 0: return `${num + 1}st`
+    case 1: return `${num + 1}nd`
+    case 2: return `${num + 1}rd`
+    default: return `${num + 1}th`
+  }
+}
+
 const ThirdStep = (props) => {
-  const { previousStep, showSuccess, submitting, invalid, ...rest } = props
+  const { previousStep, showSuccess, recoveryPhrase, submitting, invalid, ...rest } = props
   const { indexes, handleSubmit, isMnemonicVerified } = rest
 
   return (
@@ -45,19 +83,26 @@ const ThirdStep = (props) => {
       </SuccessOverlay>
       <VerificationContainer authType={showSuccess && isMnemonicVerified}>
         <Container>
-          {indexes.map(index => <WordInput index={index} />)}
+          {indexes.map(index =>
+            <WordContainer>
+              <Text size='13px' weight={300}>
+                {`${languageHelper(index)} word`}
+              </Text>
+              <Field name={`word${index}`} component={TextBox} validate={[required]} center />
+            </WordContainer>
+          )}
         </Container>
         <Buttons>
           <Button type='submit' nature='primary' disabled={submitting || invalid}>
             <FormattedMessage id='scenes.securitycenter.walletrecoveryphrase.thirdstep.confirm' defaultMessage='Confirm' />
           </Button>
-          {
-            invalid
-              ? <Link size='12px' weight={300} onClick={previousStep}>
-                <FormattedMessage id='scenes.securitycenter.walletrecoveryphrase.thirdstep.doublecheck' defaultMessage='Double check your backup phrase' />
-              </Link>
-              : null
-          }
+          {/*{*/}
+            {/*invalid*/}
+              {/*? <Link size='12px' weight={300} onClick={previousStep}>*/}
+                {/*<FormattedMessage id='scenes.securitycenter.walletrecoveryphrase.thirdstep.doublecheck' defaultMessage='Double check your backup phrase' />*/}
+              {/*</Link>*/}
+              {/*: null*/}
+          {/*}*/}
         </Buttons>
       </VerificationContainer>
     </Form>
